+++ lld/COFF/SymbolTable.h
@@ -160,6 +160,7 @@ public:
 
   void fixupExports();
   void assignExportOrdinals();
+  void parseModuleDefs(StringRef path);
 
   // Iterates symbols in non-determinstic hash table order.
   template <typename T> void forEachSymbol(T callback) {